backend/crawlers/flashscore.py
"""
FlashScore crawler implementation

NOTE: This is a TEMPLATE implementation.
FlashScore has anti-bot protection and requires more sophisticated techniques:
- Rotating proxies
- Browser fingerprinting bypass
- Cloudflare bypass
- Session management

For production, consider:
1. Using official APIs (if available)
2. Using paid proxy services (ScraperAPI, Bright Data)
3. Using more sophisticated anti-detection tools
"""
from typing import Dict, List, Optional
from bs4 import BeautifulSoup
import structlog

from crawlers.base import BaseCrawler

logger = structlog.get_logger()


class FlashScoreCrawler(BaseCrawler):
    """
    FlashScore crawler for live scores and match data

    WARNING: This is a simplified template.
    Production implementation requires:
    - Proxy rotation
    - Anti-bot bypass
    - Rate limiting
    - Error handling
    """

    BASE_URL = "https://www.flashscore.com"

    async def crawl_match(self, match_id: str) -> Optional[Dict]:
        """
        Crawl single match data from FlashScore

        Args:
            match_id: FlashScore match ID

        Returns:
            Match data dict with score, events, stats
        """
        url = f"{self.BASE_URL}/match/{match_id}"

        try:
            # Use Playwright for JS-rendered content
            html = await self.fetch_html(url, use_playwright=True)

            if not html:
                logger.warning("no_html_content", url=url)
                return None

            soup = BeautifulSoup(html, 'lxml')

            # Parse match data
            # NOTE: Selectors will change! This is just a template
            match_data = {
                'score': self._parse_score(soup),
                'minute': self._parse_minute(soup),
                'status': self._parse_status(soup),
                'events': self._parse_events(soup),
                'statistics': self._parse_statistics(soup),
            }

            logger.info("match_crawled", match_id=match_id)

            return match_data

        except Exception as e:
            logger.error("crawl_match_failed", match_id=match_id, error=str(e))
            return None

    async def crawl_match_events(self, match_id: str) -> List[Dict]:
        """
        Crawl match events (goals, cards, substitutions)

        Args:
            match_id: FlashScore match ID

        Returns:
            List of event dicts
        """
        url = f"{self.BASE_URL}/match/{match_id}"

        try:
            html = await self.fetch_html(url, use_playwright=True)

            if not html:
                return []

            soup = BeautifulSoup(html, 'lxml')

            events = self._parse_events(soup)

            return events

        except Exception as e:
            logger.error("crawl_events_failed", match_id=match_id, error=str(e))
            return []

    async def crawl_fixtures(self, days_ahead: int = 7) -> List[Dict]:
        """
        Crawl upcoming fixtures

        Args:
            days_ahead: Number of days to crawl

        Returns:
            List of fixture dicts
        """
        # TODO: Implement fixture crawling
        logger.warning("crawl_fixtures_not_implemented")
        return []

    async def crawl_league_table(self, league_id: str) -> Optional[Dict]:
        """
        Crawl league standings

        Args:
            league_id: League slug (e.g., 'premier-league')

        Returns:
            League table dict
        """
        # TODO: Implement league table crawling
        logger.warning("crawl_league_table_not_implemented")
        return None

    # Private parsing methods

    def _parse_score(self, soup: BeautifulSoup) -> Dict:
        """Parse score from soup"""
        try:
            # NOTE: This is a template - selectors WILL differ
            home_score = soup.select_one('.home-score')
            away_score = soup.select_one('.away-score')

            if home_score and away_score:
                return {
                    'home': int(home_score.text.strip()),
                    'away': int(away_score.text.strip())
                }
        except:
            pass

        return {'home': 0, 'away': 0}

    def _parse_minute(self, soup: BeautifulSoup) -> Optional[int]:
        """Parse current minute"""
        try:
            minute_elem = soup.select_one('.minute')
            if minute_elem:
                return self.parse_minute(minute_elem.text)
        except:
            pass

        return None

    def _parse_status(self, soup: BeautifulSoup) -> str:
        """Parse match status"""
        try:
            status_elem = soup.select_one('.status')
            if status_elem:
                status_text = status_elem.text.strip().lower()

                if 'live' in status_text or "'" in status_text:
                    return 'live'
                elif 'finished' in status_text or 'ft' in status_text:
                    return 'finished'
                elif 'half' in status_text or 'ht' in status_text:
                    return 'halftime'

        except:
            pass

        return 'scheduled'

    def _parse_events(self, soup: BeautifulSoup) -> List[Dict]:
        """Parse match events"""
        events = []

        try:
            # NOTE: Template selectors
            event_elements = soup.select('.event-row')

            for elem in event_elements:
                event_type = self._get_event_type(elem)
                minute = self._get_event_minute(elem)
                player = self._get_event_player(elem)
                team = self._get_event_team(elem)

                if event_type and minute and player:
                    events.append({
                        'type': event_type,
                        'minute': minute,
                        'player': player,
                        'team': team,
                        'description': None
                    })

        except Exception as e:
            logger.error("parse_events_failed", error=str(e))

        return events

    def _parse_statistics(self, soup: BeautifulSoup) -> Dict:
        """Parse match statistics"""
        try:
            # NOTE: Template implementation
            return {
                'possession': {'home': 50, 'away': 50},
                'shots': {'home': 0, 'away': 0},
                'shots_on_target': {'home': 0, 'away': 0},
                'corners': {'home': 0, 'away': 0},
                'fouls': {'home': 0, 'away': 0},
            }
        except:
            return {}

    def _get_event_type(self, elem) -> Optional[str]:
        """Get event type from element"""
        # Template
        return 'goal'

    def _get_event_minute(self, elem) -> Optional[int]:
        """Get event minute from element"""
        # Template
        return 45

    def _get_event_player(self, elem) -> Optional[str]:
        """Get player name from element"""
        # Template
        return "Player Name"

    def _get_event_team(self, elem) -> str:
        """Get team (home/away) from element"""
        # Template
        return "home"
