backend/requirements.txt
# FastAPI & ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Celery & Task Queue
celery==5.3.4
celery[redis]==5.3.4

# Web Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
lxml==4.9.3
playwright==1.40.0
httpx==0.25.2
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilities
structlog==23.2.0
python-dateutil==2.8.2
pytz==2023.3

# Monitoring
prometheus-client==0.19.0

# Rate Limiting
slowapi==0.1.9

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
